# Conflicts:
#	tests/scikit_learn/etl/test_select.py
#	tests/scikit_learn/etl/test_split.py
#	tests/scikit_learn/util.py

+++ juicer/kb8s/__init__.py
@@ -1,9 +1,11 @@
 # coding=utf-8
 import json
+import datetime
 import logging.config
 
 from kubernetes import client
 from kubernetes.client.rest import ApiException
+from gettext import gettext
 
 logging.config.fileConfig('logging_config.ini')
 log = logging.getLogger('juicer.kb8s')
@@ -11,7 +13,6 @@ log = logging.getLogger('juicer.kb8s')
 
 def eval_and_kill_pending_jobs(cluster, timeout=60 * 5):
     configuration = client.Configuration()
-    configuration.host = cluster['address']
     configuration.verify_ssl = False
     configuration.debug = False
     if 'general_parameters' not in cluster:
@@ -21,7 +22,10 @@ def eval_and_kill_pending_jobs(cluster, timeout=60 * 5):
     for parameter in cluster['general_parameters'].split(','):
         key, value = parameter.split('=')
         if key.startswith('kubernetes'):
-            cluster_params[key] = value
+            if key == 'kubernetes.apiserver':
+                configuration.host = value
+            else:
+                cluster_params[key] = value
 
     token = cluster['auth_token']
     configuration.api_key = {"authorization": "Bearer " + token}
@@ -39,9 +43,7 @@ def eval_and_kill_pending_jobs(cluster, timeout=60 * 5):
 
 
 def delete_kb8s_job(workflow_id, cluster):
-    return
     configuration = client.Configuration()
-    configuration.host = cluster['address']
     configuration.verify_ssl = False
     configuration.debug = False
     if 'general_parameters' not in cluster:
@@ -51,40 +53,53 @@ def delete_kb8s_job(workflow_id, cluster):
     for parameter in cluster['general_parameters'].split(','):
         key, value = parameter.split('=')
         if key.startswith('kubernetes'):
-            cluster_params[key] = value
-
+            if key == 'kubernetes.apiserver':
+                configuration.host = value
+            else:
+                cluster_params[key] = value
     token = cluster['auth_token']
     configuration.api_key = {"authorization": "Bearer " + token}
     # noinspection PyUnresolvedReferences
     client.Configuration.set_default(configuration)
 
-    name = 'job-{}'.format(workflow_id)
+    name = 'job-{}-{}'.format(workflow_id, 
+            round(datetime.datetime.now().timestamp()))
     namespace = cluster_params['kubernetes.namespace']
 
     api = client.ApiClient(configuration)
     batch_api = client.BatchV1Api(api)
+    
+    all_jobs = batch_api.list_namespaced_job(namespace=namespace, watch=False)
+    body = client.V1DeleteOptions(propagation_policy='Background')
+    for k8s_job in all_jobs.items:
+        name = k8s_job.metadata.name
+        if name.startswith('job-{}-'.format(workflow_id)):
+            try:
+                log.info('Deleting Kubernetes job %s', name)
 
-    try:
-        log.info('Deleting Kubernetes job %s', name)
-        batch_api.delete_namespaced_job(
-            name, namespace, grace_period_seconds=10, pretty=True)
-    except ApiException as e:
-        print("Exception when calling BatchV1Api->: {}\n".format(e))
+                batch_api.delete_namespaced_job(
+                    name, namespace, body=body, grace_period_seconds=1, pretty=True)
+            except ApiException as e:
+                print("Exception when calling BatchV1Api->: {}\n".format(e))
 
 
 def create_kb8s_job(workflow_id, minion_cmd, cluster):
     configuration = client.Configuration()
-    configuration.host = cluster['address']
     configuration.verify_ssl = False
     configuration.debug = False
     if 'general_parameters' not in cluster:
-        raise ValueError('Incorrect cluster config.')
+        raise ValueError(
+            gettext(
+                'Incorrect cluster config (missing general_parameters).'))
 
     cluster_params = {}
     for parameter in cluster['general_parameters'].split(','):
         key, value = parameter.split('=')
         if key.startswith('kubernetes'):
-            cluster_params[key] = value
+            if key == 'kubernetes.apiserver':
+                configuration.host = value
+            else:
+                cluster_params[key] = value
     env_vars = {
         'HADOOP_CONF_DIR': '/usr/local/juicer/conf',
     }
@@ -95,18 +110,13 @@ def create_kb8s_job(workflow_id, minion_cmd, cluster):
     client.Configuration.set_default(configuration)
 
     job = client.V1Job(api_version="batch/v1", kind="Job")
-    name = 'job-{}'.format(workflow_id)
+
+    name = 'job-{}-{}'.format(workflow_id, 
+            round(datetime.datetime.now().timestamp()))
     container_name = 'juicer-job'
     container_image = cluster_params['kubernetes.container']
     namespace = cluster_params['kubernetes.namespace']
     pull_policy = cluster_params.get('kubernetes.pull_policy', 'Always')
-
-    gpus = int(cluster_params.get('kubernetes.resources.gpus', 0))
-
-    print('-' * 30)
-    print('GPU KB8s specification: ' + str(gpus))
-    print('-' * 30)
-    log.info('GPU specification: %s', gpus)
 
     job.metadata = client.V1ObjectMeta(namespace=namespace, name=name)
     job.status = client.V1JobStatus()
@@ -135,10 +145,19 @@ def create_kb8s_job(workflow_id, minion_cmd, cluster):
     pvc_claim = client.V1PersistentVolumeClaimVolumeSource(
         claim_name='hdfs-pvc')
 
-    if gpus:
-        resources = {'limits': {'nvidia.com/gpu': gpus}}
-    else:
-        resources = {}
+    resources_params = [
+        ('gpus', int, 'nvidia.com/gpu'), 
+        ('cpu', str, 'cpu'), 
+        ('memory', str, 'memory')
+    ]
+
+    resources = {'limits': {}}
+    for resource_name, transform, kb8s_name in resources_params:
+        if resource_name in cluster_params:
+            resource_value = cluster_params.get(
+                'kubernetes.resources.' + resource_name)
+            if resource_value:
+                resources['limits'][kb8s_name] = transform(resource_value)
 
     container = client.V1Container(name=container_name,
                                    image=container_image,
